+++ packages/core/src/Node.ts
@@ -6,13 +6,13 @@ import utils from 'evm-lite-utils';
 import Account from './Account';
 import Transaction from './Transaction';
 
-export default class Node<TConsensus extends IAbstractConsensus> {
+export default class Node<TConsensus extends IAbstractConsensus | undefined> {
 	// a node requires an underlying consensus protocol (solo | babble | ...)
-	public readonly consensus: TConsensus;
+	public readonly consensus?: TConsensus;
 
 	private readonly client: Client;
 
-	constructor(host: string, port: number = 8080, consensus: TConsensus) {
+	constructor(host: string, port: number = 8080, consensus?: TConsensus) {
 		this.client = new Client(host, port);
 
 		this.consensus = consensus;