+++ packages/core/src/index.ts
@@ -5,3 +5,5 @@ export { default as Account, IAccount } from './Account';
 export { default as Contract, IAbstractSchema } from './Contract';
 
 export default Node;
+
+export { default as Monet } from './Monet';